ml_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0 